requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0

# Type checking and development
typing-extensions>=4.0.0
mypy>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.0.0 